Rename requests proto module as account.
