+++ proto/account.proto
@@ -1,4 +1,4 @@
-// Descriptions for requests.
+// Descriptors for account messages.
 
 syntax = "proto2";
 package auvsi_suas.proto;
